requirements/base.txt
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-environ==0.11.2
psycopg2-binary==2.9.9
cryptography==42.0.8
django-cors-headers==4.3.1
django-ratelimit==4.1.0
pyotp==2.9.0
